// app/model/IPlanet.ts
import { IDetails } from "./IDetails";

export interface IPlanet extends IDetails {
  name?: string;
  rotation_period?: string;
  orbital_period?: string;
  diameter?: string;
  climate?: string;
  gravity?: string;
  terrain?: string;
  surface_water?: string;
  population?: string;
  residents?: string[];
  films?: string[];
}